LayoutTests/inspector/unit-tests/wrapped-promise-expected.txt

Testing methods of WrappedPromise.


== Running test suite: WrappedPromise
-- Running test case: WrappedPromise.constructor resolve
-- Running test case: WrappedPromise.constructor reject
-- Running test case: WrappedPromise.prototype.resolve
-- Running test case: WrappedPromise.prototype.reject
-- Running test case: WrappedPromise.prototype.get settled
PASS: Should not be settled yet.
PASS: Should be settled now.
